Improve resource deletion
The issue is that resources hold a reference to the device, which get released
when the resource reaches 0 references. The resource could be the last thing
holding a reference to the device, meaning the device can get deleted as the
resource is being released. Implicit resources (swapchain 0, and auto depth-
stencil buffer) would then go away during the mParent->Release() call, so
'this' becomes invalid, while explicit resources would no longer have a valid
command queue or GL context to delete its GL resources with.

So, we have to check mIsAuto (and delete 'this' as needed) before attempting
to release the parent device.