Parse vertex shaders using mojoshader
This version of mojoshader is altered for my needs. In particular, the GLSL
uniforms are placed into blocks so that they can be bound to uniform buffers
(to share D3D's constants across programs). A future change will be to apply
the projection matrix to the final vertex output so it can be properly flipped
for offscreen rendering, have the Z rescaled for proper clipping, and apply a
half-pixel offset to match D3D's pixel centers.