Set the vertex shader program, and use generic vertex attributes as needed
Problem: glVertexAttribPointer apparently requires use a VBOs. This is not good
since I'm unsure of how to ensure the buffer is mapped for inevitable Lock
calls, without stalling in the Lock function (without shadowing the buffer and
doubling memory usage, anyway). ARB_buffer_storage provides for persistently-
mapped buffers which solves the problem perfectly, however it's not core until
GL 4.4, and the extension may not be available on the target hardware/drivers.