Use renderbuffers instead of the window's provided surfaces
This generally makes things easier in regards to matching up the various D3D
surfaces for rendering (e.g. using the swapchain's backbuffer mixed with other
color surfaces). This also simplifies viewport handling, and projection fixup
since it will always be flipped for offscreen rendering.

It does, however, incur an extra FBO blit to Present the backbuffer; it needs
to be blit to the real backbuffer, which is then swapped to the front buffer
(which itself is possibly a copy operation). But considering the other copies
that would be needed to properly handle mixing the default buffers with app-
created ones, this is likely better overall.