Implement occlusion queries
A couple caveats:

No protection against starting an occlusion query with a second query object
(this doesn't work with GL; don't know if D3D9 allows it or what it does).

Result availability (and the result data) isn't checked until GetData is
called, which then has to be checked by the background thread. so it will
return S_FALSE for a time even after GL has  the data available.