Handle special NULL format surfaces and textures
This format is a D3D9 hack which is used to make a handle to an object which
has no storage (VRAM or otherwise), and all reads/writes to it are ignored. Its
main purpose is to provide a way to disable render target 0, since the D3D9 API
doesn't allow setting a NULL surface for RT 0.