Don't wake the command queue thread for every send
The should improve performance by not having the background thread constantly
waking up and going to sleep, which can otherwise cause scheduling problems
between all the threads. The queue is instead left to its own devices and only
explicitly woken up on EndScene, after all drawing commands have been issued
(thus allowing the background thread to run uninterrupted and concurrent with
other tasks the app may be doing), and on Present, when it's time to be shown
on screen.

It is also woken up when D3DGETDATA_FLUSH is used, and any time a function call
needs to wait for a command to complete, although these should be rare during
normal play.